+++ Changelog.md
@@ -14,6 +14,10 @@
 - Add expression: listsort[list list, string expression]
 - Add expression: tostring[object]
 - Fix issue when comparing worldobjects of different instances but same Id (fixes listcontains when the list is wobjects as well)
+- Add expression: delayexec[number delay, string expression]
+- Add expression: if[object valuetocheck, string trueexpression, string falseexpression]
+- Add command: /ub id
+- Fix /ub use|portal|select all take ids in decimal or hex format on top of string names. ie /ub portal 0x733D902D
 
 ## 0.2.5 (2021-10-06) [UtilityBeltInstaller-0.2.5.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/dd7285b144ef9327bf8041c64841c22f/UtilityBeltInstaller-0.2.5.exe)
 - Add expressions: wobjectgetdoubleprop[], wobjectgetboolprop[], wobjectgetstringprop[], wobjecthasdata[], wobjectrequestdata[]